+++ jdecode_test.go
@@ -1,6 +1,7 @@
 package junosdecode_test
 
 import (
+	"fmt"
 	"math/rand"
 	"testing"
 
@@ -14,6 +15,8 @@ const (
 
 // TestDecode decode example password.
 func TestDecodePassword(t *testing.T) {
+	t.Parallel()
+
 	passwordDecoded, err := junosdecode.Decode(junWordCoded)
 	if err != nil {
 		t.Errorf("error on decode %v", err)
@@ -25,6 +28,8 @@ func TestDecodePassword(t *testing.T) {
 
 // TestDecodeBadEncoded try decode encrypted secret but with missing or extra character.
 func TestDecodeBadEncoded(t *testing.T) {
+	t.Parallel()
+
 	// remove last character
 	if passwordDecoded, err := junosdecode.Decode(junWordCoded[:len(junWordCoded)-1]); err == nil {
 		t.Errorf("missing character in junWordCoded not detected, passwordDecoded: %s", passwordDecoded)
@@ -37,5 +42,24 @@ func TestDecodeBadEncoded(t *testing.T) {
 	if passwordDecoded, err := junosdecode.Decode(
 		junWordCoded + string(rune(rand.Intn(26))) + string(rune(rand.Intn(26)))); err == nil { //nolint: gosec
 		t.Errorf("extra characters in junWordCoded not detected, passwordDecoded: %s", passwordDecoded)
+	}
+}
+
+func TestDecodeParrallel(t *testing.T) {
+	t.Parallel()
+
+	for i, v := range []string{junWordCoded, junWordCoded, junWordCoded} {
+		word := v
+		t.Run(fmt.Sprintf("Loop number %d", i), func(t *testing.T) {
+			t.Parallel()
+
+			output, err := junosdecode.Decode(word)
+			if err != nil {
+				t.Errorf("got unexpected error: %s", err)
+			}
+			if output != junWordDecoded {
+				t.Errorf("got unexpected decoded word: %s", output)
+			}
+		})
 	}
 }